requirements.txt
# UI
streamlit>=1.30

# PDF processing
pdfplumber>=0.10

# LLM
anthropic>=0.40

# Data processing
pandas>=2.1
pydantic>=2.5

# Units
pint>=0.23

# String matching
rapidfuzz>=3.5

# Text normalization
unidecode>=1.3

# CLI
click>=8.1

# Config
python-dotenv>=1.0

# Testing
pytest>=8.0
pytest-cov>=4.1
